Replace mock data with real PubMed API integration
- Remove all hardcoded mock data responses
- Implement PubMed E-utilities API integration for real publication data
- Add searchPubMed() function to query PubMed database directly
- Add formatArticlesResponse() to format results consistently
- Ensure PMIDs and DOIs now match actual publications
- Maintain fallback to OpenAI API when OPENAI_API_KEY is available
- Provide accurate, real-time publication data instead of static mock responses
